+++ lib/src/components/body/body_component.dart
@@ -22,4 +22,15 @@ class NgdDataTableBodyComponent {
 
   @Input()
   List<dynamic> data;
+
+  String getStyles(NgdDataColumn column){
+    var _styles = '';
+    if(column.width != null){
+      _styles += ' flex: 0; width: ' + column.width.toString() + 'px';
+    }
+    else if(column.flexWidth != null){
+      _styles += ' flex: ' + column.flexWidth.toString();
+    }
+    return _styles;
+  }
 }