+++ lib/src/components/header/header_component.dart
@@ -45,7 +45,10 @@ class NgdDataTableHeaderComponent {
   String getStyles(NgdDataColumn column){
     var _styles = '';
     if(column.width != null){
-      _styles += ' width: ' + column.width;
+      _styles += ' flex: 0; width: ' + column.width.toString() + 'px';
+    }
+    else if(column.flexWidth != null){
+      _styles += ' flex: ' + column.flexWidth.toString();
     }
     return _styles;
   }